requirements.txt
psycopg2-binary>=2.9.9
pymongo>=4.6.0
pandas>=2.0.0
numpy>=1.24.0
pyyaml>=6.0
python-dotenv>=1.0.0
